refactored Connection back to a module
We can tell a server and a client apart in an incoming connection by simply checking against our server list, because a server will never connect to us unless we initiated a connection previously, so if the host is unknown, we know it must be a client. This allows to eliminate an intermediary layer.